archieve/requirements.txt
langchain
psycopg2-binary
python-dotenv
openai
google-generativeai
supabase
tiktoken
unstructured
ssl
nltk
PyPDF2
atlassian-python-api
bs4
kuzu
google-cloud-aiplatform
zep-python
duckduckgo-search
nomic
gpt4all
pyllamacpp
google-search-results
langchain.text_splitter
langchain.embeddings
langchain.llms
langchain.vectorstores
langchain.chains
langchain.memory
langchain_core.messages
 
torch  
sentence-transformers  
tqdm

requests

numpy

sentencepiece

langchain

pyllamacpp
gpt4all
streamlit

unstructured
chromadb 
transformers

rich

pyfiglet
ean-text
Unidecode
nest_asyncio
